wip:exception

+++ src/main/java/mgsystem/management_system/product/ProductController.java
@@ -1,5 +1,7 @@
 package mgsystem.management_system.product;
 
+import mgsystem.management_system.error.*;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,66 +9,78 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.util.List;
+import java.util.Objects;
 
 // Handle requests from the client side.
 @Controller
 //@RestController
-//@RequestMapping("/api")
+@RequestMapping("/api")
 public class ProductController {
+    final String DIRECT_URL = "redirect:/api/index";
+    // final String DIRECT_URL = "redirect:/";
+
     @Autowired
     private ProductService service;
 
     // direct to index.html.
-    @RequestMapping(path = "/", method = RequestMethod.GET)
+//    @RequestMapping(path = "/index", method = RequestMethod.GET)
+    @GetMapping("/index")
     public String index(Model model){
-       List<Product> list = service.listAllData();
-       model.addAttribute("listProducts", list);
+//       List<Product> list = service.listAllData();
+        List<Product> list = service.AllData();
+        model.addAttribute("listProducts", list);
 
-       return "/index";
+        return "/index";
     }
 
     // direct to add product page.
-    @RequestMapping(path = "/data", method = RequestMethod.GET)
+    @GetMapping("/data")
     public String showAddPage(Model model){
         Product product = new Product();
         model.addAttribute("product",product);
         return "add_product";
     }
 
-    // direct to edit page and save the edited data.
-//    @RequestMapping(value = "/edit/{id}", method=RequestMethod.GET)
-//    public ModelAndView showEditPageAndSave(@PathVariable(name = "id") String id){
-//        int convertType = Integer.parseInt(id);
-//
-//        ModelAndView edit  = new ModelAndView("edit_product");
-//        String get = service.getOneData(convertType);
-//        edit.addObject("product",get);
-//
-//        return edit;
-//    }
-
     // insert
-    @RequestMapping(value = "/save", method = RequestMethod.POST)
+    @PostMapping("/save")
     public String saveData(@ModelAttribute("product") Product product) {
-        if(product.getName() != null && product.getPrice() > 0){
-            service.save(product);
-            return "redirect:/";
+        try{
+            if(product.getPrice() == 0 || product.getName() == null){
+//                throw new ArithmeticException();
+                HandleError.throwInputsError();
+            }else{
+                service.save(product);
+                System.out.println("成功!!");
+            }
+        }catch(ArithmeticException e){
+            System.out.println("Error: " + e);
         }
-        return "redirect:/";
+        return DIRECT_URL;
     }
 
-//    @RequestMapping(value = "/data/{id}", method = RequestMethod.GET)
-    @GetMapping(value = "/update/{id}")
-    public ModelAndView showUpdatePage(@PathVariable(name = "id") Long id){
 
-//        String getData = service.getOneData(id.intValue());
-        System.out.println("showUpdatePage");
-        Product getData = service.get(id);
-//        System.out.println(getData);
+    @GetMapping("/update/{id}")
+    public String showUpdatePage(@PathVariable(name = "id") Long id,Model model){
+        try{
+            String find = service.getOneData(id);
+            if(find.isEmpty()){
+//                throw new ArithmeticException();
+                HandleError.throwIdError();
+            }else{
+                System.out.println("showUpdatePage");
+                Product getData = service.get(id);
 
-        ModelAndView mav  = new ModelAndView("edit_product");
-        mav.addObject("product",getData);
-        return  mav;
+//                ModelAndView mav  = new ModelAndView("edit_product");
+//                mav.addObject("product",getData);
+
+                model.addAttribute("product",getData);
+//                return mav;
+            }
+        }catch(ArithmeticException e){
+            System.out.println("Error: " + e);
+        }
+
+        return "edit_product";
 
 //        model.addAttribute("product",getData);
 //        System.out.println(model);
@@ -74,20 +88,42 @@ public class ProductController {
     }
 
     // Put method
-    @RequestMapping(value="/update/{id}", method = RequestMethod.PUT)
+    @PutMapping("/update/{id}")
     public String updateData(@PathVariable("id") Long id, @RequestBody Product product){
-        String productName = product.getName();
-        Integer productPrice = product.getPrice();
+        System.out.println("updated");
+
+
+//        String productName = product.getName();
+//        Integer productPrice = product.getPrice();
 //        System.out.println(productPrice.getClass().getName());
 
-        service.update(id,productName,productPrice);
-        return "redirect:/";
+//        service.update(id,productName,productPrice);
+
+        // ---------
+        String res = service.getOneData(id);
+
+        product.setName(product.getName());
+        product.setPrice(product.getPrice());
+//        service.update(product);
+        service.save(product);
+
+        return  DIRECT_URL;
     }
 
 //     Delete
-    @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
+    @GetMapping("/delete/{id}")
     public String deleteData(@PathVariable("id") Long id){
-        service.delete(id);
-        return "redirect:/";
+        try{
+            String find = service.getOneData(id);
+            if(find.isEmpty()){
+                //                throw new ArithmeticException();
+                HandleError.throwIdError();
+            }else{
+                service.delete(id);
+            }
+        }catch(ArithmeticException e){
+            System.out.println("Error: " + e);
+        }
+        return DIRECT_URL;
     }
 }
